backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.1
python-dotenv==1.0.1
langchain==0.2.16
langchain-openai==0.1.23
